Finished piece download process

+++ SharpTorrent/P2P/Bitfield.cs
@@ -2,17 +2,17 @@ namespace SharpTorrent.P2P;
 
 public static class Bitfield
 {
-    public static bool HasPiece(byte[] bitfield, int bitIndex)
+    public static bool HasPiece(byte[] bitfield, uint bitIndex)
     {
         var byteIndex = bitIndex / 8;
         var offset = bitIndex % 8;
-        return (bitfield[byteIndex] >> (7 - offset) & 1) == 1;
+        return (bitfield[byteIndex] >> (int)(7 - offset) & 1) == 1;
     }
 
-    public static void SetPiece(byte[] bitfield, int bitIndex)
+    public static void SetPiece(byte[] bitfield, uint bitIndex)
     {
         var byteIndex = bitIndex / 8;
         var offset = bitIndex % 8;
-        bitfield[byteIndex] |= (byte)(1 << (7 - offset));
+        bitfield[byteIndex] |= (byte)(1 << (int)(7 - offset));
     }
 }
